refactor: 캐러셀 로직 분리
- 캐러셀 로직을 공용으로 사용하기 위해 custom hook으로 분리
